lint + clean
